requirements.txt
SQLAlchemy==1.3.24
gino==1.0.1
aiogram==2.22.1
asyncpg==0.26.0
python-decouple==3.6
pandas==1.4.4
openpyxl